Enable memcache stats check and flush
